pyproject.toml
[project]
name = "firstagent-onp-dataanalysis"
version = "0.1.0"
description = "On-prem data analysis agent powered by Claude Agent SDK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.0.1",
    "pandas>=2.2",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "xlsxwriter>=3.2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
